% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download RedCap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{A character string specifying the name of the API token stored in the
user's .REnviron file.}

\item{redcapUrl}{A character string specifying the URL of the RedCap server.}

\item{redcapReportId}{A character string specifying the unique ID of the RedCap report to be downloaded.}
}
\value{
A tibble containing the contents of the downloaded RedCap report.
}
\description{
This function queries the RedCap API to download a specified report using the provided API token
and URL. The report is identified by its unique report ID. The function returns the contents of
the report as a tibble.
}
\examples{
# Download RedCap report using token from .REnviron file
redcapToken = "REDCAP_API_TOKEN"
redcapUrl = "https://redcap.example.com/api/"
redcapReportId = "123456"
report_data = downloadRedcapReport(redcapToken, redcapUrl, redcapReportId)
head(report_data)

}
