% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the parameter p that maximizes the log-likelihood of a Bernoulli distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of binary outcomes (0 and 1).}
}
\value{
The p value that maximizes and the log-likelihood of the vector of binary outcomes.
}
\description{
This function takes a vector of binary outcomes 'data' and performs a grid-based search to find
the parameter p that maximizes the log-likelihood of observing the data under a Bernoulli
distribution.
}
\examples{
data = c(1, 0, 0, 0, 1, 1, 1)
logLikBernoulli(data)

}
