% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Minimum Sample Size for T-Test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numerical vector representing the first sample.}

\item{x2}{A numerical vector representing the second sample (optional).}
}
\value{
The minimum sample size needed for the t-test.
}
\description{
This function takes in either one or 2 numerical vectors, calculates the minimum sample size needed for a t-test of the null hypotheses
that either mu_x1 == 0 for one sample input or mu_x1 == mu_x2 for two samples with 80\% power at alpha=0.05. It is a wrapper around
the pwr::pwr.t2n.test function.
}
\examples{
# Calculate minimum sample size for one-sample t-test
x1 = rnorm(50, mean = 0, sd = 1)
min_n = minimumN(x1)

# Calculate minimum sample size for two-sample t-test
x1 = rnorm(50, mean = 0, sd = 1)
x2 = rnorm(60, mean = 0.5, sd = 1)
min_n = minimumN(x1, x2)

}
