% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Principal Component Analysis (PCA) Approximation}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A matrix or data frame containing the original data.}

\item{npc}{An integer specifying the number of principal components to use for the approximation.}
}
\value{
A matrix containing the approximation of the original data based on the specified number of principal components.
}
\description{
This function performs Principal Component Analysis (PCA) on the input data \code{x}, extracts the specified number of principal components 'npc', and uses them to reconstruct an approximation of the original data. The approximation is then rescaled and centered to match the scale and centering of the original data.
}
\examples{
# Example usage:
# Generate example data
set.seed(123)
x = matrix(rnorm(100), ncol = 10)
# Perform PCA approximation with 3 principal components
approx_data = pcApprox(x, npc = 3)

}
