% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Variable Names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble data frame.}
}
\value{
A tibble data frame with variable names standardized to "small_camel" case.
}
\description{
This function converts the variable names in input tibble 'data' to "small_camel" case
using the dplyr::rename_with and janitor::make_clean_names functions.
}
\examples{
data = tibble::tribble(
  ~First_Name, ~Last_Name,
  "John", "Doe"
)
standardized_data = standardizeNames(data)
# Output:
#   firstName lastName
#   <chr>     <chr>
# 1 John      Doe

}
