% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and Plot Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numerical vector indicating the status of each individual (0 = observation censored, 1 = event observed).}

\item{time}{A numerical vector indicating the time at which each event occurred or each individual was censored.}
}
\value{
A numerical vector representing the survival probabilities at each unique time point.
}
\description{
This function takes a numerical vector of status and a numerical vector of time, creates a data frame with columns \code{time} and \code{status}, sorts it by time in ascending order,
removes censored data, calculates survival probabilities, and plots a survival curve using the \code{plot} function.
}
\examples{
# Generate example data
status = c(1, 1, 0, 0, 1)
time = c(10, 15, 20, 25, 30)

# Plot survival curve
survCurv(status, time)

}
