% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Reverse the Scaling/Centering of a Vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A vector that has been scaled and/or centered.}
}
\value{
A vector with the original scale and centering reversed.
}
\description{
If the input vector \code{x} has been scaled and centered using the \code{\link{scale}} function, this function will reverse the centering and scaling to return the original unscaled vector. If the input vector has not been scaled or centered, it will be returned as-is.
}
\examples{
# Example usage:
scaled_vector = scale(1:10)
unscaled_vector = unscale(scaled_vector)

}
